Remove sudo from build script.
The "sudo" command in build.sh could make it failed, when running on a
system without this command. This issue is more frequently occured in
container, which has root permission and don't has "sudo" installed by
default.

So, it's better to remove it from build.sh and let user decide when to
add it for install.

Signed-off-by: Han Qiang <qiang.han@intel.com>