ptp: replace E2E mode with P2P mode.
End to End means the latency through the network is calculated directly
between master and slave. This allows for PTP to operate over networks
with Transparent clocks, non-PTP aware network switches, or a combination
of both.  The slaves sends  delay requests upstream to the master, the
master responds, and the network latency is calculated. The calculation
is then applied by the slave to the received Sync message.

Peer to Peer operates by calculating the delay between egress of the
upstream node to the ingress of the downstream node (i.e. the delay down
the wire)  rather than the whole network at once. As sync, announce and
follow-up messages are sent down the network, residence delay is accounted
for in each node through the correction field, and downstream devices
compensate for the delay down the wire. The difference is with the requests
and response messages, which are only sent between two connected devices
and are not used for calculating the delay past those devices.

Peer to peer is more efficient and is less affected by asymmetry within
the network. When multiple slaves are present in a network, peer delay
packets are only sent to the nearest node, not all the way upstream to
the master. Lastly, if the network is to suddenly change due to a fault
or similar, recovery time is reduced as not all delay calculations will
be affected.