ecosystem/ffmpeg: improve unicast initial (#917) (#931)
Address an issue where the RX is sending
ARP replies to the TX while still being completely uninitialized, which
leads to the session dropping frames. The problem worsens as the TX
outperforms
the RX in terms of initialization speed.

Resolve this by halting the FFmpeg plugin
initialization and starting the scheduled tasklets because we don't want
the CNI thread to respond to the ARP requests sent in the tv_attach
function
until the RX is ready to receive those packets.