Fix error handling in parsing sample freq in FFmpeg audio plugins
Commit 065923f3f768 ("Refactor FFmpeg audio plugin sample rate parsing")
added a helper function for parsing the sample frequency. However, the error
returned from the function was checked incorrectly (inverted). This led to
a segmentation fault when starting FFmpeg with MTL audio plugins.

This commit fixes the error handling by checking if the return value is non zero.

Fixes: 065923f3f768 ("Refactor FFmpeg audio plugin sample rate parsing")
Signed-off-by: Konstantin Ilichev <konstantin.ilichev@intel.com>
Change-type: DefectResolution