Fix: flow rules failing after verification on PF (#1021)
DPDK backend flow rules are failing on the E810 PF when ETH rules are
ignored on the default attribute group 0.

Flow rules are used to filter the data coming from queues, and by
default, attribute group 0 should be used. Applying the filters should not fail
when the flow validation function passes. However, this does not happen when we
use filters with the ETH segment set to NULL on E810 physical cards (it
works fine on the virtual NICs hosted on the same NIC).

Fix this problem for now by reapplying the flow rules when the
verification fails in this specific scenario.

This code should be deleted when the E810 PF NIC stops failing to apply
flow rules in this specific scenario.