Add: prototype of src rx st20 plugin for gstreamer
Create a plugin for st20p rx standard
for gstreamer, allowing to receive the data in the
pipeline mode.

Create a simple design where we create
the session in the _start function, and
the recive and processing is in _create function.

Move some of the functions from the TX RX to the
common file.
Change the name of TX plugin to mtl_st20p_tx to
make it easier to distiguish between the plugins.