Fix: pipeline mode API parameters check
Pipeline API for st20, st22, and st30 TX will
segfault when NULL parameters are passed to them.

Fix by adding a null guard that will return NULL
if any of the parameters are null

TODO: Tests for this cases and testing environment
for experimental st40 API as no testing is
implemented yet for this pipeline.