Fix: RxTxApp Meson configuration issue (#1065)
The image was not building correctly when
the bare metal directory was built before due
to the .dockerignore copying the build directory
of RxTxApp from a new location.

Fix by adding the necessary entries to
.dockerignore to prevent the build directory
from being copied to image.

Add copy of the RxTxApp to the final stage and
adjust the README to reflect the changes.

Fixes: 1580c7ff04a93664bf3751b5b533ccf23b940da1