ADD: Create versions.env with basic variables (#1156)
ADD: Create `versions.env` with basic variables regarding latest
recommended versions, that include the following:

- DPDK driver version and link
- ICE driver version and link
- IRDMA driver version and link

To use it, user can source it in his shell session and refer to params
by name or use it as a reference for manual download of dependencies.

From now on every version of packages that are mandatory for this
repository should be defined and updated in this file.