Fix: GStreamer plugins wrong memory buffer size (#1167)
In the st20tx and st30tx GStreamer plugins, the memory
buffer size was incorrectly taken from the
GstBuffer size which includes the combined size
of all GstMemory elements in the buffer, as we
Allow multiple GstMemory elements in the buffer
This would result in a wrong buffer parsing.

Fix via taking the size from the memory map.