Add: CiCd setup script (#1154)
Add a script to set up the  environment for CI/CD that
will set up the whole project.

Add the following features to the build gh cation
* Ecosystem builds all the supported MTL plugin apps
* Build the GPU support library and mtl with GPU support
* Build the hooks for Rust and Python 
* Build the tools in /tools directory
* Build the dpdk automatically 
The purpose is to have an all-encompassing point for CICD
that can check the whole project.

Add a warning that the mediaSDK is unsupported.
Improve the reliability of existing scripts by changing
The paths. 
Delete unused variables from versions.env 
Move the common script to use versions.env 

The script is supposed to follow the "user" route, so we 
Use a few suboptimal solutions by design. 

Known Issues:
* Using git as a code downloading / patching tool
* Needing user input for the first Python installation in some cases