Add: RxTxApp user controlled pacing to tx_st20p_app
Add user controlled pacing to the RxTxApp,
this allows users to work with lipsync scnenarios
as we have global clock with the same base for all
sessions.

We jump the base time to the newest st2110 session
base time, as this was the best way of achiving
this without having to overhaul the way we initialize
The pipeline sessions.