Fix: User controlled sync for pacing st2110-20 (#1240)
A mechanism that allows the session to move
the pacing window if the current epoch is just
a little bit behind was applied to user controlled pacing,
which in turn made the scenarios where
big PTP updates or other small issues may cause
incremental delays, which is not really a problem
as long as we don't enable user controlled pacing.

Fix it by ignoring all discrepancies for user
controlled pacing scenarios (required_tai) and
just forcing the session to send the packets asap.