Fix: DPDK build script undefined behavior
The DPDK build script was using git to patch the DPDK repo inside
the MTL repository, which was causing undefined behavior.
Switch to patch to ensure this always works and add help for
UX.

Add a way of handling down ports in gtest.sh due to
e830 bug (showing up as down after reset).

Add handling to DMA vfio binds as not all DMA works
out of the box.