Fix: ST40/ST41 first packet timestamp
Packet timestamps were only calculated when
the RTP time differed from the previous frame,
which failed for the initial packet. This caused
the first TX packet to be transmitted with
timestamp 0, resulting in DDoS-like behavior
on the RX side.

During initalization make the session calculate
the timestamp to avoid such issues.

To test set the date into problematic clock area
sudo date -s "2025-10-29 17:28:30"