Add: Guard mechanism against all-redundant state
If the session ancillary seq_id or any other session timestamp
somehow moves backward in time, the MTL RX will wait indefinitely
for that threshold to be reached. While this state should never
occur under normal conditions, we can implement a threshold guard
that accepts the packet after a while.