Fix: remove rdma support (#1334)
The RDMA support will not be maintained supported the library has
a specific purpose, and RDMA is not expected to be used. This was added as
an experimental feature, so supporting this is not needed anymore.

The performance of this experimental feature was poor.
The idea of sending buffers as small as MTU via RDMA was not the brightest one.