Ci: Unify the building scripts to unlock nightly
Nighlty tests require building GStreamer and
python GStreamer tools in a consistant matter. This
commit refactors the build scripts to ensure
that they follow the same pattern and can be
easily invoked from the GitHub Actions workflow.

Make sure that tests are properly explained and
documented in the validation framework.